requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
alembic
psycopg2-binary
email-validator
jinja2
aiofiles
bcrypt
pytest>=8.2
requests>=2.32
